adds cleanup cron for dynamically generated voice channels (2)
