+++ internal/handlers/readyEvent/ready.go
@@ -17,7 +17,7 @@ func Ready(s *discordgo.Session, event *discordgo.Ready) {
 
 	utils.LogHandlerCall("Ready", "")
 
-	// Load static data once runtime is confirmed
+	// Load static data once Discord API runtime features are confirmed
 	LoadStaticData()
 
 	// Retrieve list of DB users at startup time (for convenience and some optimisation further down the line)
@@ -51,12 +51,12 @@ func Ready(s *discordgo.Session, event *discordgo.Ready) {
 	go channelHandlers.HandleExperienceGrantsMessages(false)
 	go channelHandlers.HandleChannelCreationMessages(s)
 
-	// CRON FUNCTIONS FOR VARIOUS FEATURES (like activity streaks, XP gaining?, etc.)
-	cron.ProcessUpdateActivityStreaks(24, 0, 0)               // the hh:mm:ss timestamp in a day to run the cron at
-	cron.ProcessRemoveExpiredWarns(2)                         // run every n=2 months
-	cron.ProcessClearExpiredTimeouts(s)                       // clear timeouts with freq from env var
-	cron.ProcessRemoveArchivedTimeouts(1)                     // run every n=1 month
-	cron.ProcessMonthlyLeaderboard(s, 23, 55, 00, true, true) // run on last day at given time
-	cron.ProcessCleanupUnusedDynamicChannels(s, globals.DiscordMainGuildId)
+	// CRON FUNCTIONS FOR VARIOUS FEATURES (like activity streaks, cleanups, etc.)
+	cron.ProcessUpdateActivityStreaks(24, 0, 0)                                      // the hh:mm:ss timestamp in a day to run the cron at
+	cron.ProcessRemoveExpiredWarns(2)                                                // run every n=2 months
+	cron.ProcessClearExpiredTimeouts(s)                                              // clear timeouts with freq from env var
+	cron.ProcessRemoveArchivedTimeouts(1)                                            // run every n=1 month
+	cron.ProcessMonthlyLeaderboard(s, 23, 55, 00, true, true)                        // run on last day at given time
+	cron.ProcessCleanupUnusedDynamicChannels(s, globals.DiscordMainGuildId, 60*60*5) // run every n=5 hours
 
 }
